github/requirements.txt
PyNaCl==1.5.0
requests==2.31.0
